pyproject.toml
[project]
name = "glow-worm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
    "alembic>=1.13",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "starlette-csrf>=3.0",
    "itsdangerous>=2.1",
    "pytz>=2025.2",
    "apscheduler>=3.10,<4.0",
    "fastmcp>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
